requirements.txt
Django==4.2.6
psycopg2==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
